refactor: 내비게이션 구조화
